+++ backend/discord/list.go
@@ -15,7 +15,7 @@ func (b *Bot) list(ctx context.Context, w corde.ResponseWriter, i *corde.Interac
 		user = i.Data.Resolved.Users.First()
 	}
 
-	chars, err := collection.Characters(ctx, b.Store, user.ID)
+	chars, err := collection.Characters(ctx, b.Store, uint64(user.ID))
 	if err != nil {
 		w.Respond(rspErr("An error occurred dialing the database, please try again later"))
 		return
